refactor: js mobules
